Make GStreamingExecutor abstract
