+++ modules/gapi/src/compiler/gcompiler.cpp
@@ -2,7 +2,7 @@
 // It is subject to the license terms in the LICENSE file found in the top-level directory
 // of this distribution and at http://opencv.org/license.html.
 //
-// Copyright (C) 2018-2020 Intel Corporation
+// Copyright (C) 2018-2022 Intel Corporation
 
 
 #include "precomp.hpp"
@@ -34,6 +34,7 @@
 
 #include "executor/gexecutor.hpp"
 #include "executor/gstreamingexecutor.hpp"
+#include "executor/gtbbstreamingexecutor.hpp"
 #include "backends/common/gbackend.hpp"
 #include "backends/common/gmetabackend.hpp"
 #include "backends/streaming/gstreamingbackend.hpp" // cv::gimpl::streaming::kernels()
@@ -474,8 +475,8 @@ cv::GStreamingCompiled cv::gimpl::GCompiler::produceStreamingCompiled(GPtr &&pg)
 
     GModel::ConstGraph cgr(*pg);
 
-    std::unique_ptr<GStreamingExecutor> pE(new GStreamingExecutor(std::move(pg),
-                                                                  m_args));
+    std::unique_ptr<GStreamingExecutor> pE(new GTBBStreamingExecutor(std::move(pg),
+                                                                     m_args));
     if (!m_metas.empty() && !outMetas.empty())
     {
         compiled.priv().setup(m_metas, outMetas, std::move(pE));
